Refactor alpha_bar_fn in scheduling_discrete_state.py
The commit refactors the alpha_bar_fn function in the scheduling_discrete_state.py file. The **2 exponentiation operation has been removed from the return statement of the function. This change was made by the authors.

The commit also includes some minor changes in the DiscreteStateScheduler class. The beta_start and beta_end variables have been removed, and the beta_schedule variable has been set to "linear" when the transition_mat_type is "gaussian". Additionally, the beta_schedule variable has been set to "jsd" when the transition_mat_type is "absorbing".